Refactor test organization
