+++ tests/graph/test_graph_utils.py
@@ -9,13 +9,13 @@ import pandas as pd
 import pytest
 from numpy.testing import assert_array_equal
 from pandas.testing import assert_frame_equal
-from pixelator.graph import (
-    Graph,
+from pixelator.graph.utils import (
     components_metrics,
     create_node_markers_counts,
     edgelist_metrics,
     update_edgelist_membership,
 )
+from pixelator.graph import Graph
 
 
 def add_random_names_to_vertexes(graph: Graph) -> None:
@@ -61,106 +61,6 @@ def pentagram_graph_fixture():
 def random_sequence(size: int) -> str:
     """Create a random sequence of size (size)."""
     return "".join(random.choices("CGTA", k=size))
-
-
-def test_build_graph_full_bipartite(full_graph_edgelist: pd.DataFrame):
-    """Build full-bipartite graph."""
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=True,
-    )
-    assert graph.vcount() == 50 + 50
-    assert graph.ecount() == 50 * 50
-    assert "markers" in graph.vs.attributes()
-    assert sorted(list(graph.vs[0]["markers"].keys())) == ["A", "B"]
-    assert graph.vs.attributes() == ["name", "markers", "type", "pixel_type"]
-
-
-def test_build_graph_a_node_projected(full_graph_edgelist: pd.DataFrame):
-    """Build an A-node projected graph."""
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=False,
-    )
-    assert graph.vcount() == 50
-    assert graph.ecount() == ((50 * 50) / 2) - (50 / 2)
-    assert "markers" in graph.vs.attributes()
-    assert sorted(list(graph.vs[0]["markers"].keys())) == ["A", "B"]
-    assert graph.vs.attributes() == ["name", "markers", "type", "pixel_type"]
-
-
-def test_layout_coordinates_all_pixels(full_graph_edgelist: pd.DataFrame):
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=True,
-    )
-    result = graph.layout_coordinates(only_keep_a_pixels=False)
-    assert result.shape == (100, 4)
-    assert set(result.columns) == {"x", "y", "A", "B"}
-
-
-def test_layout_coordinates_3d_layout(full_graph_edgelist: pd.DataFrame):
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=True,
-    )
-    result = graph.layout_coordinates(
-        layout_algorithm="fruchterman_reingold_3d", only_keep_a_pixels=False
-    )
-    assert set(result.columns) == {
-        "x",
-        "y",
-        "z",
-        "x_norm",
-        "y_norm",
-        "z_norm",
-        "A",
-        "B",
-    }
-    assert result.shape == (100, 8)
-
-
-def test_layout_coordinates_only_a_pixels(full_graph_edgelist: pd.DataFrame):
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=True,
-    )
-    result = graph.layout_coordinates(only_keep_a_pixels=True)
-    assert result.shape == (50, 4)
-    assert set(result.columns) == {"x", "y", "A", "B"}
-
-
-def test_layout_coordinates_3d_layout_only_a_pixels(full_graph_edgelist: pd.DataFrame):
-    graph = Graph.from_edgelist(
-        edgelist=full_graph_edgelist,
-        add_marker_counts=True,
-        simplify=True,
-        use_full_bipartite=True,
-    )
-    result = graph.layout_coordinates(
-        layout_algorithm="fruchterman_reingold_3d", only_keep_a_pixels=True
-    )
-    assert set(result.columns) == {
-        "x",
-        "y",
-        "z",
-        "x_norm",
-        "y_norm",
-        "z_norm",
-        "A",
-        "B",
-    }
-    assert result.shape == (50, 8)
 
 
 def test_components_metrics(full_graph_edgelist: pd.DataFrame):
